feat(note-form): add react-hook-form on note-from
Also, separete the note form and transcribe dialog in different components to be more readable and maintainable.
