+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@formkit/auto-animate": "^0.8.1",
+    "@hookform/resolvers": "^3.3.4",
     "@radix-ui/react-alert-dialog": "^1.0.5",
     "@radix-ui/react-collapsible": "^1.0.3",
     "@radix-ui/react-dialog": "^1.0.5",
@@ -35,6 +36,7 @@
     "next-themes": "^0.2.1",
     "react": "link:@formkit/auto-animate/react",
     "react-dom": "^18",
+    "react-hook-form": "^7.51.2",
     "sonner": "^1.4.3",
     "tailwind-merge": "^2.2.1",
     "zod": "^3.22.4"
